py/pyproject.toml
[project]
name = "nba-hack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyter>=1.1.1",
    "pandas>=2.2.3",
    "pbpstats>=1.3.11",
    "torch>=2.6.0",
    "torch-geometric>=2.6.1",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "openai>=1.13.0",
    "python-dotenv>=1.0.0",
]
